arp_app/backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
websockets>=12.0
scapy>=2.5.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
joblib>=1.3.0
pyyaml>=6.0.1
python-multipart>=0.0.6
pydantic>=2.5.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.0
psutil>=5.9.6
aiofiles>=23.2.1
sqlalchemy>=2.0.23
alembic>=1.12.1 